+++ CHANGELOG.md
@@ -15,4 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added RAG augmentation ([11dfb8b](https://github.com/ydethe/ragwebui/commit/11dfb8b6f3ea04ec3dfd6bfa4d5246f5ee84f3b1) by Yann de The).
 - Added API keys and remove of deprecated method QdrantClient.search ([7a56da9](https://github.com/ydethe/ragwebui/commit/7a56da902ec471077d7c8d5dd4ab53b000160de6) by Yann de The).
 
+### Fixed
+
+- Fixing citations ([a0e4da6](https://github.com/ydethe/ragwebui/commit/a0e4da64794eabaaee921bc56abd10527bd325d5) by Yann de The).
+
 <!-- insertion marker -->